250416/09-01-DRF/requirements.txt
asgiref==3.8.1
Django==4.2.20
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.13.2
